refactor: change templates path
