+++ src/components/templates/FullScreenButtonTemplate.tsx
