+++ src/components/templates/FullScreenCenterTemplate.tsx
