+++ src/stories/templates/FullScreenCenterTemplate.stories.tsx
@@ -1,9 +1,9 @@
 import MobileScreen from 'components/common/MobileScreen';
 import ScreenSpinner from 'components/common/ScreenSpinner';
-import FullScreenCenterTemplate from 'components/common/templates/FullScreenCenterTemplate';
+import FullScreenCenterTemplate from 'components/templates/FullScreenCenterTemplate';
 
 export default {
-  title: 'common/templates/Full Screen Center Template',
+  title: 'templates/Full Screen Center Template',
   component: FullScreenCenterTemplate,
 };
